Add a tile layer from GSI

Add a tile layer from Geospatial Information Authority of Japan (GSI).

gsi_tiles

Format

A 48 length, leaflet objects.

Details

Stores map tiles that can be used with leaflets. Please follow the terms and conditions of use for the applicable tile at http://maps.gsi.go.jp/development/ichiran.html when using it. It contains tiles that can be used as base maps for interactive maps based on leaflet. See example section its use in leaflet. To use a mapview, a tile name is given to mapview::mapview(map = ).

Examples

names(gsi_tiles)

#>  [1] "standard"             "pale"                 "english"             
#>  [4] "lcm25k_2012"          "lcm25k"               "ccm1"                
#>  [7] "ccm2"                 "vbm"                  "vbmd_bm"             
#> [10] "vbmd_colorrel"        "vbmd_pm"              "vlcd"                
#> [13] "lum200k"              "lake1"                "lakedata"            
#> [16] "blank"                "seamlessphoto"        "ortho"               
#> [19] "airphoto"             "gazo4"                "gazo3"               
#> [22] "gazo2"                "gazo1"                "ort_old10"           
#> [25] "ort_USA10"            "ort_riku10"           "relief"              
#> [28] "anaglyphmap_color"    "anaglyphmap_gray"     "hillshademap"        
#> [31] "earthhillshade"       "slopemap"             "slopezone1map"       
#> [34] "afm"                  "lcmfc2"               "lcmfc1"              
#> [37] "swale"                "cp"                   "jikizu2015_chijiki_d"
#> [40] "jikizu2015_chijiki_i" "jikizu2015_chijiki_f" "jikizu2015_chijiki_h"
#> [43] "jikizu2015_chijiki_z" "jikizu_chijikid"      "jikizu_chijikii"     
#> [46] "jikizu_chijikif"      "jikizu_chijikih"      "jikizu_chijikiz"     

require("leaflet")

#> Loading required package: leaflet

gsi_tiles[[1]]

gsi_tiles[[1]] %>%
  addCircles(
    data = sf::st_transform(extreme_points %>%
      purrr::reduce(c),
      crs = 4326))