Clip and move some geometries for mapping

Move geometry differently from the real-life arrangement for mapping. When displaying a map showing Japan, the southern islands are sometimes moved. To achieve this, we need to perform false operations on the geometry.

move_jpn_rs(data, clip = TRUE)

Arguments

data
sf that records the prefecture or municipality of Japan

clip
An option to hide isolated island that are separated from other geometry and have a small area.

Examples

require("sf")
move_jpn_rs(jgd2011_bbox)

#> Simple feature collection with 17 features and 2 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 128.1335 ymin: 30.22511 xmax: 148.8954 ymax: 45.55733
#> CRS:            EPSG:6668
#> First 10 features:
#>    system epsg                       geometry
#> 1       I 6669 MULTIPOLYGON (((133.6426 42...
#> 2      II 6670 MULTIPOLYGON (((130.9688 30...
#> 3     III 6671 MULTIPOLYGON (((132.0229 33...
#> 4      IV 6672 MULTIPOLYGON (((132.4816 32...
#> 5       V 6673 MULTIPOLYGON (((134.8116 34...
#> 6      VI 6674 MULTIPOLYGON (((135.817 33....
#> 7     VII 6675 MULTIPOLYGON (((137.0057 34...
#> 8    VIII 6676 MULTIPOLYGON (((138.219 34....
#> 9      IX 6677 MULTIPOLYGON (((139.9176 31...
#> 10      X 6678 MULTIPOLYGON (((141.6014 38...