Clip and move some geometries for mapping

Move geometry differently from the real-life arrangement for mapping. When displaying a map showing Japan, the southern islands are sometimes moved. To achieve this, we need to perform false operations on the geometry.

move_jpn_rs(data, clip = TRUE)

Arguments

data
sf that records the prefecture or municipality of Japan

clip
An option to hide isolated island that are separated from other geometry and have a small area.

Examples

require("sf")
move_jpn_rs(jgd2011_bbox)

#> Simple feature collection with 17 features and 2 fields
#> geometry type:  MULTIPOLYGON
#> dimension:      XY
#> bbox:           xmin: 128.1335 ymin: 30.22511 xmax: 148.8954 ymax: 45.55733
#> CRS:            EPSG:6668
#> # A tibble: 17 x 3
#>    system epsg                                                          geometry
#>    <fct>  <fct>                                               <MULTIPOLYGON [°]>
#>  1 I      6669  (((133.6426 42.0186, 133.6409 42.0206, 133.6353 42.02073, 133.6…
#>  2 II     6670  (((130.9688 30.41987, 130.9683 30.41987, 130.9682 30.4206, 130.…
#>  3 III    6671  (((132.0229 33.7352, 132.0246 33.73687, 132.0305 33.74014, 132.…
#>  4 IV     6672  (((132.4816 32.79633, 132.4794 32.7978, 132.481 32.80053, 132.4…
#>  5 V      6673  (((134.8116 34.15647, 134.8115 34.16293, 134.813 34.1674, 134.8…
#>  6 VI     6674  (((135.817 33.44587, 135.8139 33.44753, 135.8133 33.44833, 135.…
#>  7 VII    6675  (((137.0057 34.66687, 137.003 34.66733, 137.0021 34.66987, 136.…
#>  8 VIII   6676  (((138.219 34.6184, 138.2134 34.62087, 138.211 34.61947, 138.21…
#>  9 IX     6677  (((139.9176 31.8869, 139.9177 31.8868, 139.9178 31.88653, 139.9…
#> 10 X      6678  (((141.6014 38.3856, 141.6019 38.3872, 141.6018 38.38927, 141.6…
#> 11 XI     6679  (((139.7997 41.35347, 139.8 41.3584, 139.7945 41.35833, 139.798…
#> 12 XII    6680  (((143.2568 41.91574, 143.2577 41.91598, 143.258 41.9152, 143.2…
#> 13 XIII   6681  (((145.9984 43.36213, 145.9968 43.36567, 145.9927 43.36787, 145…
#> 14 XIV    6682  (((142.9601 30.22516, 142.9569 30.22511, 142.9543 30.22616, 142…
#> 15 XV     6683  (((133.0835 41.15127, 133.0848 41.1534, 133.0832 41.15393, 133.…
#> 16 XVI    6684  (((128.9908 39.04473, 128.9822 39.0458, 128.9763 39.04627, 128.…
#> 17 XVII   6685  (((136.3869 39.46169, 136.3847 39.46343, 136.3835 39.46611, 136…